Updates to the SymbolStore
