+++ src/main/scala/cromwell/engine/db/RealDataAccess.scala
@@ -3,7 +3,7 @@ package cromwell.engine.db
 import java.util.Date
 
 import cromwell.binding.Call
-import cromwell.engine.SymbolStore.SymbolStoreEntry
+import cromwell.engine.store.{SymbolStoreEntry, SymbolStore}
 import cromwell.engine.{WorkflowId, WorkflowState}
 
 // TODO Insert real data access logic, rename.
