+++ src/test/scala/cromwell/HelloWorldActorSpec.scala
@@ -10,7 +10,8 @@ import cromwell.HelloWorldActorSpec._
 import cromwell.binding.values.WdlString
 import cromwell.binding.{WorkflowDescriptor, WorkflowOutputs, UnsatisfiedInputsException, WdlNamespace}
 import cromwell.engine._
-import cromwell.engine.WorkflowActor._
+import cromwell.engine.workflow.WorkflowActor
+import WorkflowActor._
 import cromwell.engine.backend.local.LocalBackend
 import cromwell.util.SampleWdl.HelloWorld
 import cromwell.util.SampleWdl.HelloWorld.Addressee
