+++ src/test/scala/cromwell/ThreeStepActorSpec.scala
@@ -9,7 +9,8 @@ import com.typesafe.config.ConfigFactory
 import cromwell.CromwellSpec.DockerTest
 import cromwell.binding._
 import cromwell.binding.values.{WdlInteger, WdlValue}
-import cromwell.engine.WorkflowActor._
+import cromwell.engine.workflow.WorkflowActor
+import WorkflowActor._
 import cromwell.engine._
 import cromwell.engine.backend.local.LocalBackend
